+++ theme/gatsby-config.js
@@ -5,6 +5,11 @@ module.exports = themeOptions => {
     siteMetadata: {
       title: 'Doctocat',
       shortName: 'Doctocat',
+      header: {
+        title: 'Primer',
+        url: 'https://primer.style',
+        logoUrl: 'https://primer.style'
+      },
       description: 'A Gatsby theme for creating Primer documentation sites',
       imageUrl: 'https://user-images.githubusercontent.com/10384315/53922681-2f6d3100-402a-11e9-9719-5d1811c8110a.png'
     },