+++ theme/gatsby-node.js
@@ -44,7 +44,7 @@ exports.createPages = async ({graphql, actions}, themeOptions) => {
       const rootAbsolutePath = path.resolve(process.cwd(), themeOptions.repoRootPath || '.')
 
       const fileRelativePath = path.relative(rootAbsolutePath, node.fileAbsolutePath)
-      const defaultBranch = themeOptions.defaultBranch || 'master'
+      const defaultBranch = themeOptions.defaultBranch || 'main'
       const editUrl = getEditUrl(repo, fileRelativePath, defaultBranch)
 
       let contributors = []
@@ -86,6 +86,7 @@ exports.onPostBuild = async ({graphql}) => {
               frontmatter {
                 componentId
                 status
+                a11yReviewed
               }
             }
           }
@@ -97,7 +98,8 @@ exports.onPostBuild = async ({graphql}) => {
       return {
         id: node.context.frontmatter.componentId,
         path: node.path,
-        status: node.context.frontmatter.status.toLowerCase()
+        status: node.context.frontmatter.status.toLowerCase(),
+        a11yReviewed: node.context.frontmatter.a11yReviewed || false
       }
     })