+++ theme/src/components/rails-link.js
@@ -1,5 +1,5 @@
 import {Link, Text, Box} from '@primer/react'
-import {BookIcon} from '@primer/octicons-react'
+import {RubyIcon} from '@primer/octicons-react'
 import React from 'react'
 
 function SourceLink({href}) {
@@ -7,8 +7,8 @@ function SourceLink({href}) {
     <li>
       <Link href={href} target="_blank">
         <Box sx={{display: 'flex', gap: 2, alignItems: 'center'}}>
-          <BookIcon />
-          <Text>Storybook</Text>
+          <RubyIcon />
+          <Text>Rails</Text>
         </Box>
       </Link>
     </li>